# Improve performance of lib_oval.py 's getElementById method