This came up at https://github.com/nexB/vulnerablecode/pull/179 as we wanted to make parser for OVAL documents to obtain mappings of vulnerabilities and packages. 

I have used  https://people.canonical.com/~ubuntu-security/oval/com.ubuntu.xenial.cve.oval.xml.bz2 which is about 40 MB to obtain these below stats
Before performance tweaks :  
```

...: p = pstats.Stats('perf.txt') 
   ...: p.sort_stats('cumulative').print_stats(10)                                                                           
Sat May 16 15:19:18 2020    perf.txt

         769968949 function calls (769934359 primitive calls) in 383.945 seconds

   Ordered by: cumulative time
   List reduced from 469 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     27/1    0.000    0.000  383.945  383.945 {built-in method builtins.exec}
        1    0.000    0.000  383.945  383.945 profiler.py:1(<module>)
        1    0.434    0.434  382.571  382.571 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/oval_parser.py:22(get_data)
    84661   12.240    0.000  205.191    0.002 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/lib_oval.py:408(getElementByID)
    33373    9.235    0.000  195.203    0.006 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/oval_parser.py:115(get_object_state_of_test)
    15267   59.637    0.004  162.846    0.011 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/oval_parser.py:95(get_tests_of_definition)
289921428   75.677    0.000  123.842    0.000 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/lib_oval.py:734(getId)
    33373    2.101    0.000  112.006    0.003 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/lib_oval.py:356(getStates)
    33373   52.569    0.002  109.479    0.003 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/lib_oval.py:377(<listcomp>)
110130900   56.909    0.000   56.909    0.000 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/lib_oval.py:1358(__init__)
```
As you can see the old getElementById method was the culprit.
The new getElementById method(the one included in the PR)  gives these stats 

```
Sat May 16 17:18:32 2020    perf3.txt

         4094392 function calls (4059802 primitive calls) in 5.144 seconds

   Ordered by: cumulative time
   List reduced from 476 to 10 due to restriction <10>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     27/1    0.000    0.000    5.144    5.144 {built-in method builtins.exec}
        1    0.000    0.000    5.144    5.144 profiler.py:1(<module>)
        1    0.140    0.140    3.727    3.727 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/oval_parser.py:60(get_data)
    33373    0.072    0.000    1.665    0.000 /home/shivam/coding/opensource/oriv/vulnerablecode/vulnerabilities/scraper/oval_parser.py:173(get_versionsrngs_from_state)
    33371    0.110    0.000    1.581    0.000 /home/shivam/coding/opensource/oriv/vulnerablecode/venv/lib/python3.8/site-packages/dephell_specifier/range_specifier.py:20(__init__)
        1    0.000    0.000    1.325    1.325 /usr/lib/python3.8/xml/etree/ElementTree.py:1192(parse)
        1    0.000    0.000    1.325    1.325 /usr/lib/python3.8/xml/etree/ElementTree.py:571(parse)
        1    1.325    1.325    1.325    1.325 {method '_parse_whole' of 'xml.etree.ElementTree.XMLParser' objects}
    33371    0.107    0.000    1.300    0.000 /home/shivam/coding/opensource/oriv/vulnerablecode/venv/lib/python3.8/site-packages/dephell_specifier/range_specifier.py:44(_parse)
    33382    0.152    0.000    0.808    0.000 /home/shivam/coding/opensource/oriv/vulnerablecode/venv/lib/python3.8/site-packages/dephell_specifier/specifier.py:51(__init__)
```
As you can see the new `getElementById` doesn't even make to top 10 time consuming function. 

**How this is acheived** 
The old `getElementById ` works something as , take an OVAL id, check every element for it's id ,if this matches to the requested id, return this element. **Old time complexity O(N)**.

The new `getElementById` instead, maintains  a dictionary [id->element]  and uses it. **New time complexity O(1)**.
Signed-off-by: Shivam Sandbhor <shivam.sandbhor@gmail.com>